+++ src/services/Settings.vala
@@ -33,6 +33,7 @@ public class Terminal.Settings : PQMarble.Settings {
   public bool    easy_copy_paste                      { get; set; }
   public bool    fill_tabs                            { get; set; }
   public bool    headerbar_drag_area                  { get; set; }
+  public bool    notify_process_completion            { get; set; }
   public bool    pretty                               { get; set; }
   public bool    remember_window_size                 { get; set; }
   public bool    show_headerbar                       { get; set; }